refactoring login code
