+++ src/app/[locale]/(auth)/confirm/route.ts
@@ -1,12 +1,9 @@
-// src/app/[locale]/auth/confirm/route.ts
-
 import { type EmailOtpType } from '@supabase/supabase-js'
 import { type NextRequest } from 'next/server'
 import { createClient } from '@/lib/supabase/server'
 import { redirect } from 'next/navigation'
 
 export async function GET(request: NextRequest) {
-  // Extrai os parâmetros da URL
   const searchParams = new URL(request.url).searchParams
   const token_hash = searchParams.get('token_hash')
   const type = searchParams.get('type') as EmailOtpType | null
